[wip] Move the agent class to a new file.
